+++ packages/hooks/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @reservoir0x/relay-kit-hooks
 
+## 3.0.23
+
+### Patch Changes
+
+- Updated dependencies [88a63b6]
+- Updated dependencies [88a63b6]
+  - @relayprotocol/relay-sdk@6.1.2
+
+## 3.0.22
+
+### Patch Changes
+
+- Updated dependencies [ce4d192]
+  - @relayprotocol/relay-sdk@6.1.1
+
+## 3.0.21
+
+### Patch Changes
+
+- Updated dependencies [4711f4b]
+  - @relayprotocol/relay-sdk@6.1.0
+
 ## 3.0.20
 
 ### Patch Changes