+++ packages/relay-ethers-wallet-adapter/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @reservoir0x/relay-ethers-wallet-adapter
 
+## 31.0.2
+
+### Patch Changes
+
+- Updated dependencies [88a63b6]
+- Updated dependencies [88a63b6]
+  - @relayprotocol/relay-sdk@6.1.2
+
+## 31.0.1
+
+### Patch Changes
+
+- Updated dependencies [ce4d192]
+  - @relayprotocol/relay-sdk@6.1.1
+
+## 31.0.0
+
+### Patch Changes
+
+- Updated dependencies [4711f4b]
+  - @relayprotocol/relay-sdk@6.1.0
+
 ## 30.0.1
 
 ### Patch Changes